[fix]: calendar widget 오류 수정
