+++ apps/www/src/widgets/calendar/index.ts
@@ -1,2 +1,3 @@
 export { default as CalendarWidget } from './CalendarWidget';
 export { default as CalendarWidgetConfigure } from './CalendarWidgetConfigure';
+export { default as DatePickerCalendarHeader } from './DatePickerCalendarHeader';
