+++ apps/www/src/widgets/components/EditTemplate.tsx
@@ -122,10 +122,7 @@ function EditTemplate() {
               />
             ) : null}
             {thirdModalState.items ? (
-              <CongratulationList
-                items={thirdModalState.items as OwnerAccountItem[]}
-                handleClickCopy={handleClickCopy}
-              />
+              <CongratulationList items={thirdModalState.items} handleClickCopy={handleClickCopy} />
             ) : null}
           </Modal>
         </>
