+++ apps/www/src/widgets/components/Widget.tsx
@@ -3,7 +3,7 @@
 import type { IInvitation, WidgetItem } from '@repo/shared/src/types/pageBrothers.type';
 import dynamic from 'next/dynamic';
 import { memo } from 'react';
-import { useFieldArray, useWatch } from 'react-hook-form';
+import { useWatch } from 'react-hook-form';
 
 import { useWidgetIndex } from '../hooks';
 import type { HookFormValues } from '../types';
@@ -51,22 +51,17 @@ function UnmemoizedWidget({ invitation, widgetItem, isMultiModal }: WidgetProps)
   const WidgetComponent = components[widgetItem.type as keyof typeof components];
   const widgetIndex = useWidgetIndex(widgetItem)!;
 
-  const { fields } = useFieldArray<HookFormValues, `invitation.widgets`>({
-    name: `invitation.widgets` as const,
-  });
+  const changingInvitation =
+    useWatch<HookFormValues>({
+      name: 'invitation',
+    }) ?? invitation;
 
-  const changingInvitation = useWatch<HookFormValues>({
-    name: 'invitation',
-  });
-
-  // 최초 청첩장 생성시 인덱스가 없어서 인트로 위젯 미리보기 관련 에러가 발생
-  // 따라서 인덱스가 없으면 최초 위젯 아이템을 사용
-  const widgetItemItem = fields.filter((field) => field.index === widgetIndex)[0] ?? widgetItem;
+  const changingWidgetItem = (changingInvitation as IInvitation).widgets[widgetIndex] ?? widgetItem;
 
   return (
     <WidgetComponent
       invitation={changingInvitation ? (changingInvitation as IInvitation) : invitation}
-      widgetItem={widgetItemItem}
+      widgetItem={changingWidgetItem as WidgetItem}
       isMultiModal={isMultiModal}
     />
   );
