+++ apps/www/src/widgets/hooks.tsx
@@ -1,13 +1,14 @@
 'use client';
 
-import type { WidgetItem } from '@repo/shared';
+import type { IInvitation, WidgetItem } from '@repo/shared';
+import { format, isSameDay } from 'date-fns';
 import type { ChangeEvent, Dispatch, SetStateAction } from 'react';
-import { createContext, useCallback, useContext, useMemo, useState } from 'react';
+import { createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 
 import useModalStore from '@/www/widgets/zustand';
 
 import WidgetCombobox from './components/WidgetCombobox';
-import { getWidgetIndex } from './utils';
+import { getCombinedDateTime, getWidgetIndex } from './utils';
 
 export function useWidgetIndex(widgetItem: WidgetItem | Omit<WidgetItem, 'id'>) {
   const { invitation } = useModalStore();
@@ -110,4 +111,91 @@ export function InputErrorProvider({ children }: { children: React.ReactNode }) 
       {children}
     </InputErrorContext.Provider>
   );
+}
+
+interface UseCalendarSelectionProps {
+  invitation: IInvitation;
+  setValue: (name: string, value: string) => void;
+}
+
+export function useCalendarSelection({ invitation, setValue }: UseCalendarSelectionProps) {
+  const [selectedDay, setSelectedDay] = useState<Date | undefined>(new Date(invitation.eventAt));
+  const [selectedTime, setSelectedTime] = useState<Date | undefined>(new Date(invitation.eventAt));
+  const [isTimeClicked, setIsTimeClicked] = useState<boolean>(false);
+
+  const { closeMultiModal } = useModalStore();
+
+  const handleOnChange = (date: Date | null) => {
+    if (date) setSelectedDay(date);
+  };
+
+  const handleOnChangeTime = (date: Date | null) => {
+    if (date) setSelectedTime(date);
+    setIsTimeClicked(true);
+  };
+
+  const originDate = useMemo(() => {
+    return getCombinedDateTime(new Date(invitation.eventAt), new Date(invitation.eventAt));
+  }, [invitation]);
+
+  const combinedDateTime = useMemo(() => {
+    if (!selectedDay || !selectedTime) return;
+    return getCombinedDateTime(selectedDay, selectedTime);
+  }, [selectedDay, selectedTime]);
+
+  useEffect(() => {
+    const selectedTimeToCenter = document.querySelector(
+      '.react-datepicker__time-list-item--selected',
+    );
+    if (selectedTimeToCenter) {
+      selectedTimeToCenter.scrollIntoView({
+        behavior: 'instant',
+        block: 'center',
+        inline: 'center',
+      });
+    }
+  }, [selectedTime]);
+
+  useEffect(() => {
+    if (selectedDay && selectedTime && combinedDateTime) {
+      const datesAreEqual = isSameDay(originDate, combinedDateTime);
+      if (!datesAreEqual || !isTimeClicked) {
+        const formatted = format(combinedDateTime, "yyyy-MM-dd'T'HH:mm:ss");
+        setValue('invitation.eventAt', formatted);
+        if (isTimeClicked) {
+          closeMultiModal();
+          return;
+        }
+      }
+      if (isTimeClicked) {
+        closeMultiModal();
+      }
+    }
+  }, [
+    selectedDay,
+    selectedTime,
+    setValue,
+    closeMultiModal,
+    isTimeClicked,
+    originDate,
+    combinedDateTime,
+    invitation,
+  ]);
+
+  // useEffect(() => {
+  //   return () => {
+  //     if (!combinedDateTime) return;
+  //     const formatted = format(combinedDateTime, "yyyy-MM-dd'T'HH:mm:ss");
+  //     setValue('invitation.eventAt', formatted);
+  //   };
+  // }, [setValue, isTimeClicked, combinedDateTime]);
+
+  return {
+    selectedDay,
+    selectedTime,
+    handleOnChange,
+    handleOnChangeTime,
+    originDate,
+    combinedDateTime,
+  };
 }
