+++ apps/www/src/widgets/intro/IntroWidget.tsx
@@ -19,7 +19,6 @@ function IntroWidgetComp({
   invitation,
   isMultiModal = false,
 }: IntroWidgetProps): React.ReactNode {
-  console.log('IntroWidgetComp', widgetItem);
   return (
     <WidgetWrapper widgetItem={widgetItem} isMultiModal={isMultiModal}>
       <div className="no-interaction">
